+++ io-type/core/src/commonMain/kotlin/com/wsr/conv/DotExt.kt
@@ -2,27 +2,27 @@ package com.wsr.conv
 
 import com.wsr.BLAS
 
-infix fun Array<DoubleArray>.dot(other: Array<DoubleArray>): Array<DoubleArray> {
+infix fun Array<FloatArray>.dot(other: Array<FloatArray>): Array<FloatArray> {
     val m = other.size
     val n = size
     val k = this[0].size
 
     // 入力行列を1次元配列に変換 (row-major)
-    val aFlat = DoubleArray(n * k)
+    val aFlat = FloatArray(n * k)
     for (i in 0 until n) {
         for (j in 0 until k) {
             aFlat[i * k + j] = this[i][j]
         }
     }
 
-    val bFlat = DoubleArray(m * k)
+    val bFlat = FloatArray(m * k)
     for (i in 0 until m) {
         for (j in 0 until k) {
             bFlat[i * k + j] = other[i][j]
         }
     }
 
-    val cFlat = DoubleArray(m * n)
+    val cFlat = FloatArray(m * n)
 
     // C[m, n] = A[n, k]^T * B[m, k]^T = B[m, k] * A[n, k]^T
     // other[f] dot this[i] = sum_k(other[f][k] * this[i][k])
@@ -33,18 +33,18 @@ infix fun Array<DoubleArray>.dot(other: Array<DoubleArray>): Array<DoubleArray> 
         m = m,
         n = n,
         k = k,
-        alpha = 1.0,
+        alpha = 1f,
         a = bFlat,
         lda = k,
         b = aFlat,
         ldb = k,
-        beta = 0.0,
+        beta = 0f,
         c = cFlat,
         ldc = n,
     )
 
     // 結果を2次元配列に変換
-    val result = Array(m) { DoubleArray(n) }
+    val result = Array(m) { FloatArray(n) }
     for (i in 0 until m) {
         for (j in 0 until n) {
             result[i][j] = cFlat[i * n + j]
@@ -54,7 +54,7 @@ infix fun Array<DoubleArray>.dot(other: Array<DoubleArray>): Array<DoubleArray> 
     return result
 }
 
-infix fun DoubleArray.dot(other: DoubleArray): Double = BLAS.ddot(
+infix fun FloatArray.dot(other: FloatArray): Float = BLAS.ddot(
     n = this.size,
     x = this,
     incX = 1,
